refactor: restructure lib and types directories, simplify exports
