+++ frontend/src/lib/trees/index.ts
@@ -1,3 +1,4 @@
 export * from './extractNearbyTrees';
 export * from './loadAdoptedTrees';
 export * from './filters';
+export * from './topSpecies';
