+++ frontend/tsconfig.json
@@ -9,11 +9,14 @@
 		"skipLibCheck": true,
 		"sourceMap": true,
 		"strict": true,
-		"moduleResolution": "bundler"
-		// "baseUrl": ".",
-		// "paths": {
-		// 	"$components/*": ["src/components/*"]
-		// }
+		"moduleResolution": "bundler",
+		"baseUrl": ".",
+		"paths": {
+			"$components/*": ["src/components/*"],
+			"$lib/*": ["src/lib/*"],
+			"$types": ["src/types"],
+			"$types/*": ["src/types/*"]
+		}
 	}
 	// Path aliases are handled by https://kit.svelte.dev/docs/configuration#alias
 	//
